Update AsyncTCP.cpp
There is a memory leak in the AsyncClient. The SSL part is not freed by the destructor. My opinion is that it is caused by _lwip_fin where we have
_pcb = NULL;
but when we want to delete ssl part by int tcp_ssl_free(struct tcp_pcb *tcp) 
it takes +pcb as an argument but it is NULL... I think this as the simplest solution.